refactor(nav): share settings shortcut handler
