+++ components/nav-user.tsx
@@ -2,8 +2,9 @@
 
 import { ChevronsUpDown, ExternalLink, Info, Settings } from 'lucide-react'
 
-import { useEffect, useState } from 'react'
+import { useCallback, useState } from 'react'
 import { ClientOnly } from '@/components/client-only'
+import { useSettingsShortcut } from '@/components/nav-user/use-settings-shortcut'
 import { SettingsDialog } from '@/components/settings'
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
 import {
@@ -39,20 +40,8 @@ export function NavUser({
 }) {
   const { isMobile } = useSidebar()
   const [settingsOpen, setSettingsOpen] = useState(false)
-
-  // Keyboard shortcut for settings (Cmd/Ctrl + ,)
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      // Cmd+, or Ctrl+, to open settings (standard for Settings/Preferences)
-      if ((event.metaKey || event.ctrlKey) && event.key === ',') {
-        event.preventDefault()
-        setSettingsOpen(true)
-      }
-    }
-
-    document.addEventListener('keydown', handleKeyDown)
-    return () => document.removeEventListener('keydown', handleKeyDown)
-  }, [])
+  const openSettings = useCallback(() => setSettingsOpen(true), [])
+  useSettingsShortcut(openSettings)
 
   // If Clerk is enabled, use Clerk navigation
   if (isClerkEnabled() && ClerkNavWrapper) {
