+++ components/nav-user/use-settings-shortcut.ts
@@ -1,0 +1,17 @@
+'use client'
+
+import { useEffect } from 'react'
+
+export function useSettingsShortcut(openSettings: () => void) {
+  useEffect(() => {
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if ((event.metaKey || event.ctrlKey) && event.key === ',') {
+        event.preventDefault()
+        openSettings()
+      }
+    }
+
+    document.addEventListener('keydown', handleKeyDown)
+    return () => document.removeEventListener('keydown', handleKeyDown)
+  }, [openSettings])
+}
